Convert docs to JSDoc style and explicitly mark public API
This begins the process of transforming existing documentation-style
comments to the JSDoc format. Mark each as being public (@public) so
that a future generation process can export only public API.

- [x] Spring
- [x] SpringConfig
- [x] SpringSystem
- [ ] Loopers
- [ ] MathUtil
- [ ] util
- [ ] BouncyConversion?
- [ ] OrigamiValueConverter
- [ ] PhysicsState

For context, I'm working on a Docusaurus-built version of the
documentation and examples in a single site. Generating docs from these
will be part of that build process.