Support using a Spring as an Observable
This adds support for using a spring as an observable.

Implemented using `Symbol.observable` and any compatible Observable
library (RxJS, etc), the movement of a spring can be observed with
`Observable.from(spring).subscribe(...)`, which returns a subscription
object that can be unsubscribed from, which internally removes the
listener from the spring.

Inside the subscription, one could -- for example -- map over the value to scale it, then
update a dom node to reflect the movement of the spring.